+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/backend/rust/RustFileWriter.java
@@ -141,6 +141,7 @@ public class RustFileWriter implements ICompilerComponent {
             "use sqllib::{\n" +
             "    *,\n" +
             "    casts::*,\n" +
+            "    binary::*,\n" +
             "    geopoint::*,\n" +
             "    timestamp::*,\n" +
             "    interval::*,\n" +