+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/ir/type/primitive/DBSPTypeBinary.java
@@ -1,0 +1,76 @@
+package org.dbsp.sqlCompiler.ir.type.primitive;
+
+import org.dbsp.sqlCompiler.compiler.frontend.CalciteObject;
+import org.dbsp.sqlCompiler.compiler.visitors.inner.InnerVisitor;
+import org.dbsp.sqlCompiler.ir.expression.literal.DBSPLiteral;
+import org.dbsp.sqlCompiler.ir.type.DBSPType;
+import org.dbsp.util.IIndentStream;
+
+import static org.dbsp.sqlCompiler.ir.type.DBSPTypeCode.BYTES;
+
+/**
+ * Represents a byte array.
+ */
+public class DBSPTypeBinary extends DBSPTypeBaseType {
+    public static final int UNLIMITED_PRECISION = -1;
+
+    public final int precision;
+
+    public DBSPTypeBinary(CalciteObject node, int precision, boolean mayBeNull) {
+        super(node, BYTES, mayBeNull);
+        this.precision = precision;
+    }
+
+    public DBSPTypeBinary(CalciteObject node, boolean mayBeNull) {
+        super(node, BYTES, mayBeNull);
+        this.precision = UNLIMITED_PRECISION;
+    }
+
+    @Override
+    public DBSPType setMayBeNull(boolean mayBeNull) {
+        if (this.mayBeNull == mayBeNull)
+            return this;
+        return new DBSPTypeBinary(this.getNode(), this.precision, mayBeNull);
+    }
+
+    @Override
+    public DBSPLiteral defaultValue() {
+        // return new DBSPByteArrayLiteral(new byte[], this.mayBeNull);
+        return null;
+    }
+
+    @Override
+    public boolean sameType(DBSPType type) {
+        DBSPTypeBinary other = type.as(DBSPTypeBinary.class);
+        if (other == null)
+            return false;
+        if (!super.sameNullability(type))
+            return false;
+        return this.precision == other.precision;
+    }
+
+    @Override
+    public boolean hasCopy() {
+        return false;
+    }
+
+    @Override
+    public void accept(InnerVisitor visitor) {
+        if (visitor.preorder(this).stop()) return;
+        visitor.push(this);
+        visitor.pop(this);
+        visitor.postorder(this);
+    }
+
+    @Override
+    public IIndentStream toString(IIndentStream builder) {
+        if (this.precision == UNLIMITED_PRECISION)
+            return super.toString(builder);
+        return builder.append(this.shortName())
+                .append("(")
+                .append(this.precision)
+                .append(")")
+                .append(this.mayBeNull ? "?" : "");
+    }
+}
+